+++ enterprise/admin/actions.py
@@ -7,6 +7,7 @@ from __future__ import absolute_import, unicode_literals
 
 import unicodecsv
 from six import string_types
+
 from django.http import HttpResponse
 from django.utils.translation import ugettext_lazy as _